Add UI, Music, New Footsteps, Faster Footsteps
I've also updated the footsteps audio engine to force a new footstep sound to play at regular intervals so it's not dependent on the length of the audio file. One trivial growing-pain of this is that because some audio files have excessive lead-in, it sounds like no footstep is being played sometimes since all there is time to hear on those ones is their silent lead-in.