add mix bus for each audio channel type adjust mixing levels add ai voice for start
I added a SFX, music, dialogue, and reverb bus for each of the types of audio to allow for smoother adjustment of Db levels. i added default_bus_layout.tres as a file for the default layout of this. I added mp3 versions of the footstep_foley_[number].mp3 and did some general mixing.